--- Entrega 3/src/compiladorImagenes/aplicacion/comandos/base_comand.py
@@ -1,14 +1,0 @@
-from functools import singledispatch
-from abc import ABC, abstractmethod
-
-class Comand:
-    ...
-
-class ComandHandler(ABC):
-    @abstractmethod
-    def handle(self, comando: Comand):
-        raise NotImplementedError()
-
-@singledispatch
-def ejecutar_commando(comando):
-    raise NotImplementedError(f'No existe implementación para el comando de tipo {type(comando).__name__}')
