--- Entrega 3/src/compiladorImagenes/aplicacion/comandos/health_check.py
@@ -1,8 +1,0 @@
-from src.commands.base_command import BaseCommand
-from flask import jsonify
-
-class HealthCheck(BaseCommand):
-  
-    def execute(self):
-        health_status = {"message": "pong"}
-        return jsonify(health_status), 200
