+++ Entrega 3/src/compiladorImagenes/seedwork/infraestructura/utils.py
@@ -1,11 +1,13 @@
 from os import getenv
 
 def broker_host():
-    return getenv('BROKER_HOST', default="amqp://guest:guest@us2.pitunnel.net:50513/")
+    return getenv(
+        "BROKER_HOST", default="http://localhost:8080"
+    )
 
 def listar_topicos():
     return {
-        'topico_entrada': "DatosDescarga",
-        'topico_salida_1': "Compilaciones",
-        'topico_salida_2': "Notificaciones"
+        'topico_entrada': "persistent://public/default/DatosDescarga",
+        'topico_salida_1': "persistent://public/default/Compilaciones",
+        'topico_salida_2': "persistent://public/default/Notificaciones"
     }
